requirements.txt
flask==3.0.3
anthropic==0.25.1
pygithub==2.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
